test(daemon): real integration test for the pairing endpoints (#994)
POST /api/pair/exchange is security-critical — anonymous, rate-limited,
issues bearer tokens, enforces a per-IP failure lockout — but had only
fake-reimplementation tests, the same gap that hid the reminder regression.

- Extract the three pairing endpoints (POST /api/pair/exchange,
  GET/DELETE /api/pair/devices) from inline Program.cs handlers into a
  MapPairingEndpoints extension method (Netclaw.Daemon/Security/
  PairingEndpointRouteBuilderExtensions.cs), matching MapWebhookEndpoints /
  MapReminderEndpoints — pure mechanical move, no logic change.
- Replace the fake-reimplementation PairingExchangeEndpointTests and
  PairingCodeEndpointTests with PairingEndpointRouteBuilderExtensionsTests,
  which calls the real app.MapPairingEndpoints() with real PairingCodeService,
  PairingExchangeGuard, and DeviceRegistry.
- Covers: no pending code (404), valid code (200 + token, anonymous, device
  registered), invalid code (401), guard lockout (429 + Retry-After), missing
  fields (400), duplicate device (409), consumed/expired code, the issued
  token authenticating a later request, and the authenticated device-list /
  revoke endpoints.
- Reinstates the two reverse-proxy tests the extraction would otherwise have
  dropped: behind a trusted proxy, both the per-IP lockout and the rate
  limiter must partition by the X-Forwarded-For client IP, not the proxy's.

Build clean, 0 warnings, slopwatch 0 issues, headers verified, Daemon.Tests
537 green.
